third_party/WebKit/LayoutTests/fast/css/invalidation/targeted-class-content-pseudo-expected.txt

CONSOLE WARNING: line 25: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Check that targeted class invalidation works for ::content selectors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(r1, null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 5
PASS getComputedStyle(r1, null).backgroundColor is green
PASS getComputedStyle(r2, null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 3
PASS getComputedStyle(r2, null).backgroundColor is green
PASS getComputedStyle(r3, null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(r3, null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
